requirements.txt
langchain_openai
langchain_core
langchain_community
langserve
fastapi
uvicorn
python-dotenv
openai
pydantic